LayoutTests/fast/events/touch/ios/show-keyboard-after-preventing-touchstart-expected.txt


Verifies that the keyboard shows up even after preventing default on touchstart. To manually test, tap the textarea; the textarea should remain focused, and the keyboard should appear.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS keyboard was shown.
PASS document.activeElement is textarea
PASS successfullyParsed is true

TEST COMPLETE
